requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
backports.zoneinfo==0.2.1
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.114.0
greenlet==3.0.3
h11==0.14.0
idna==3.8
imbalanced-learn==0.12.3
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pathspec==0.12.1
platformdirs==4.3.2
psycopg2==2.9.9
pydantic==2.9.0
pydantic_core==2.23.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
threadpoolctl==3.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6